src/jsws.js: replace mock webserver with a real node.js webserver.
This currently returns a static string, rather than reading things
from the filesystem. The test framework also, annoyingly, prints
"Terminated" after it shuts down the web server. This will be fixed
when the web server deamonizes properly, though it might be worth
looking at a little hack to the test framework to get rid of this
in the meantime.